+++ src/scvi/model/_jaxscvi.py
@@ -163,13 +163,6 @@ class JaxSCVI(JaxTrainingMixin, BaseModelClass):
 
         return self.module.as_numpy_array(latent)
 
-    def get_normalized_expression(
-        self,
-    ):
-        # Refer to function get_accessibility_estimates
-        msg = f"get_normalized_expression is not implemented for {self.__class__.__name__}."
-        raise NotImplementedError(msg)
-
     def to_device(self, device):
         pass